+++ test/helpers/project_test.rb
@@ -5,7 +5,8 @@ describe Project do
 
   let(:project_path) { '/tmp/repos' }
   let(:path) { git_projects_path(project_path) }
-  let(:config_path) { "#{path}/git-projects.yml" }
+  let(:config_path) { File.join(path, 'git-projects.yml') }
+  let(:project_path) { '/test/tmp/projects' }
 
   before do
     clean_projects_path(project_path)
@@ -15,16 +16,36 @@ describe Project do
   end
 
   describe '#all' do
-
     context 'when size is called' do
       it 'returns correct number of projects' do
         @project.all.size.must_equal 5
       end
     end
 
+    context 'when group parameter exists' do
+
+      before do
+        @project.new_group('a','ruby')
+      end
+
+      it 'should filter projects by group' do
+        @project.all('ruby').size.must_equal 1
+      end
+    end
+  end
+
+  describe '#first' do
     context 'when there is a root_dir' do
+      it 'includes the name of the project' do
+        @project.first[0].must_equal 'a'
+      end
+
       it 'should not include the root_dir' do
         @project.first[1].keys.must_include 'root_dir'
+      end
+
+      it 'incudes the remotes' do
+        @project.first[1]['origin'].must_include project_path
       end
     end
 
@@ -35,21 +56,7 @@ describe Project do
     end
   end
 
-  describe '#first' do
-
-    it 'includes the name of the project' do
-      @project.first[0].must_equal 'a'
-    end
-
-    it 'incudes the remotes' do
-      @project.first[1]['origin'].must_include project_path
-    end
-  end
-
   describe '#set_root_path' do
-
-    let(:project_path) { '/tmp/projects' }
-
     before do
       path = Dir.pwd+project_path
       @project.set_root_path(path)
@@ -63,4 +70,27 @@ describe Project do
 
   end
 
+  describe '#new_remote' do
+    before do
+      @project.new_remote('a','bitbucket', @project.first[1]['origin'])
+    end
+
+    context 'when remote is added to an existing project' do
+      it 'adds a new remote' do
+        @project.first[1]['bitbucket'].must_equal @project.first[1]['origin']
+      end
+    end
+  end
+
+  describe '#new_group' do
+    before do
+      @project.new_group('a','ruby')
+    end
+
+    context 'when a new group is set for existing project' do
+      it 'changes the group' do
+        @project.first[1]['group'].must_equal 'ruby'
+      end
+    end
+  end
 end