+++ test/test_repos.rb
@@ -25,4 +25,30 @@ module TestRepos
     end
   end
 
+  def add_remote(project, name, url=nil)
+
+    unless File.exists?(ENV['GIT_PROJECTS'])
+      GitProject.create_config(ENV['GIT_PROJECTS'].gsub('git-projects.yml', ''))
+    end
+
+    @project = Project.new(ENV['GIT_PROJECTS'])
+    @project = @project.new_remote(project, name, @project.first[1]['origin'])
+    `rm #{ENV['GIT_PROJECTS']}` if File.exists?(ENV['GIT_PROJECTS'])
+    File.open(ENV['GIT_PROJECTS'], "w") do |f|
+      f.write @project.to_yaml.gsub(/- /,'').gsub(/    /,'  ').gsub(/---/,'')
+    end
+  end
+
+  def change_group(project, name)
+    unless File.exists?(ENV['GIT_PROJECTS'])
+      GitProject.create_config(ENV['GIT_PROJECTS'].gsub('git-projects.yml', ''))
+    end
+
+    @project = Project.new(ENV['GIT_PROJECTS'])
+    @project = @project.new_group(project, name)
+    `rm #{ENV['GIT_PROJECTS']}` if File.exists?(ENV['GIT_PROJECTS'])
+    File.open(ENV['GIT_PROJECTS'], "w") do |f|
+      f.write @project.to_yaml.gsub(/- /,'').gsub(/    /,'  ').gsub(/---/,'')
+    end
+  end
 end